+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qrgrid/styles
 
+## 1.1.0
+
+### Minor Changes
+
+- onGenerated event
+
 ## 1.0.0
 
 ### Major Changes